+++ js/main.js
@@ -1,1 +1,5 @@
-import './generator-data';
+import {photoItems} from './generat-data/createPhoto';
+
+photoItems();
+
+
